third_party/blink/web_tests/fast/forms/range/range-reset-expected.txt

Bug 15251: REGRESSION: <input type=range> doesn't respond to form.reset() or setting input.value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  
Slider with explicit limits
Current value is: "30"
PASS slider.value is "80"
Resetting form
PASS slider.value is "30"


Slider with default limits, and no explicit value
Current value is: "50"
PASS slider.value is "80"
Resetting form
PASS slider.value is "50"


Slider with default limits, and a step value
Current value is: "60"
PASS slider.value is "80"
Resetting form
PASS slider.value is "60"


PASS successfullyParsed is true

TEST COMPLETE
